# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP-STOCK-LNB API",
        "description": "Description de votre API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/immobilisations": {
            "get": {
                "tags": [
                    "Immobilisations"
                ],
                "summary": "Liste des immobilisations",
                "operationId": "4c65654d2af9005b61a847f66dc19222",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des immobilisations"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Immobilisation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Immobilisations"
                ],
                "summary": "Créer une nouvelle immobilisation",
                "operationId": "3d18686d6e5c6863bbdab8b9afe3336e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_groupe_type_immo",
                                    "id_sous_type_immo",
                                    "id_status_immo"
                                ],
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "example": "Scanner HP"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "IMMO-2025-002"
                                    },
                                    "montant_ttc": {
                                        "type": "integer",
                                        "example": 250000
                                    },
                                    "date_acquisition": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_mise_en_service": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "id_groupe_type_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_sous_type_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_status_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fournisseur_id": {
                                        "type": "integer"
                                    },
                                    "employe_id": {
                                        "type": "integer"
                                    },
                                    "bureau_id": {
                                        "type": "integer"
                                    },
                                    "vehicule_id": {
                                        "type": "integer"
                                    },
                                    "isVehicule": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Immobilisation créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceImmobilisationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/immobilisations/{id}": {
            "put": {
                "tags": [
                    "Immobilisations"
                ],
                "summary": "Mettre à jour une immobilisation",
                "operationId": "830b3ac5530f952c093862fa78258215",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'immobilisation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_groupe_type_immo",
                                    "id_sous_type_immo",
                                    "id_status_immo"
                                ],
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "example": "Scanner HP"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "IMMO-2025-002"
                                    },
                                    "montant_ttc": {
                                        "type": "integer",
                                        "example": 250000
                                    },
                                    "date_acquisition": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "date_mise_en_service": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "id_groupe_type_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_sous_type_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_status_immo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fournisseur_id": {
                                        "type": "integer"
                                    },
                                    "employe_id": {
                                        "type": "integer"
                                    },
                                    "bureau_id": {
                                        "type": "integer"
                                    },
                                    "vehicule_id": {
                                        "type": "integer"
                                    },
                                    "isVehicule": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Immobilisation mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceImmobilisationResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Immobilisations"
                ],
                "summary": "Supprimer une immobilisation",
                "operationId": "e1d120152511dc5c8740cbe16025ada9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Immobilisation supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Immobilisation supprimée avec succès"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Immobilisation": {
                "title": "Immobilisation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "designation": {
                        "type": "string",
                        "example": "Imprimante Canon"
                    },
                    "code": {
                        "type": "string",
                        "example": "IMMO-2025-001"
                    },
                    "montant_ttc": {
                        "type": "integer",
                        "example": 150000
                    },
                    "date_acquisition": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "date_mise_en_service": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-01"
                    },
                    "date_mouvement": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-10"
                    },
                    "duree_amorti": {
                        "type": "integer",
                        "example": 36
                    },
                    "duree_ammortissement": {
                        "type": "integer",
                        "example": 36
                    },
                    "taux_ammortissement": {
                        "type": "integer",
                        "example": 10
                    },
                    "etat": {
                        "type": "string",
                        "example": "Bon état"
                    },
                    "observation": {
                        "type": "string",
                        "example": "Affectée au bureau 1"
                    },
                    "isVehicule": {
                        "type": "boolean",
                        "example": false
                    },
                    "vehicule_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "id_groupe_type_immo": {
                        "type": "integer"
                    },
                    "id_sous_type_immo": {
                        "type": "integer"
                    },
                    "id_status_immo": {
                        "type": "integer"
                    },
                    "fournisseur_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "employe_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "bureau_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "bureau": {
                        "$ref": "#/components/schemas/Bureau"
                    },
                    "groupe_type_immo": {
                        "$ref": "#/components/schemas/Groupe_type_immo"
                    },
                    "sous_type_immo": {
                        "$ref": "#/components/schemas/Sous_type_immo"
                    },
                    "status_immo": {
                        "$ref": "#/components/schemas/Status_immo"
                    },
                    "employe": {
                        "$ref": "#/components/schemas/Employe"
                    },
                    "fournisseur": {
                        "$ref": "#/components/schemas/Fournisseur"
                    }
                },
                "type": "object"
            },
...
            "Bureau": {
                "title": "Bureau",
                "description": "Modèle d'un bureau",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "libelle_bureau": {
                        "type": "string",
                        "example": "Bureau Central"
                    },
                    "valeur": {
                        "type": "string",
                        "example": "Valeur optionnelle"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "Groupe_type_immo": {
                "title": "Groupe_type_immo",
                "description": "Modèle d'un Groupe_type_immo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Table Parakou"
                    },
                    "compte": {
                        "type": "integer",
                        "example": "24440005"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    }
                },
                "type": "object"
            },
            "Sous_type_immo": {
                "title": "Sous_type_immo",
                "description": "Modèle d'un Sous_type_immo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_type_immo": {
                        "type": "integer"
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Mobilier de Bureau Parakou"
                    },
                    "compte": {
                        "type": "string",
                        "example": "24440005"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-25T14:00:00Z"
                    },
                    "typeImmo": {
                        "$ref": "#/components/schemas/Bureau"
                    }
                },
                "type": "object"
            },
...
            "PostResourceImmobilisationResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Immobilisation"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}